New change for IEX handler

+++ database_objects/favorites.py
@@ -17,29 +17,39 @@ class FavoritesTable(table.DatabaseTable):
 
     def add_favorite(self, username: str, favorite: str):
         """adds a favorite for the user, returns True if successful else False"""
-        where = "Username='{}' AND Ticker='{}'".format(username, favorite)
-        # Only add favorite if no result is found
-        if self.select_from(["1"], where) == []:
-            query = "INSERT INTO {} (Username, Ticker) VALUES ('{}', '{}')".format(
-                self.table_name, username, favorite)
-            self.execute(query)
+        # abort if value is already in table
+        if favorite in self.get_favorites(username):
             return True
-        return False
+
+        # Insert value
+        query = "INSERT INTO {} (Username, Ticker) VALUES ('{}', '{}')".format(
+            self.table_name, username, favorite)
+        self.execute(query)
+        
+        # Check if value was inserted properly
+        return True if favorite in self.get_favorites(username) else False
 
     def remove_favorite(self, username: str, favorite: str):
         """removes a favorite for the user, returns True if successful else False"""
+        # check if favorite exists for user
+        if favorite not in self.get_favorites(username): 
+            return True
+
+        # Delete favorite on user
         query = "DELETE FROM {} WHERE Username ='{}' AND Ticker='{}';".format(
             self.table_name, username, favorite)
         self.execute(query)
-        return True
+
+        # Check if favorite exists on user
+        return True if favorite not in self.get_favorites(username) else False
 
     def remove_all_favorites(self, username: str):
         query = "DELETE FROM {} WHERE Username ='{}';".format(self.table_name, username)
         if self.sanitize(username):
             self.execute(query)
-        return True
+        return len(self.get_favorites(username)) == 0
 
-    def get_all_favorites(self, username: str):
+    def get_favorites(self, username: str):
         # Return all favorites of a user
         results_2D = self.select_from(["Ticker"], "Username='{}'".format(username))
         # Switch from 2-D array to 1-D
